+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "book-graveyard",
-  "version": "0.1.2",
+  "version": "0.1.3",
   "description": "",
   "main": "index.js",
   "scripts": {
